requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
gif-for-cli==1.1.2
idna==2.9
pathspec==0.8.0
Pillow==7.1.2
pkg-resources==0.0.0
python-vlc==3.0.9113
regex==2020.5.14
requests==2.23.0
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
x256==0.0.3
